Add backup.js to run MySQL / MongoDB S3 backup
This allows us to create a cronjob in Kubernetes which runs just this file, while also allowing us to skip the built-in backup cronjobs with the PREVENT_BACKUP_CRONJOBS environment variable set to `true`.

This commit also prevents the normal mysql & mongodb backups to be run when the S3 variant is turned on.